refactor: split oversized test modules for maintainability (JTN-133)
Split test_settings_blueprint.py (~1163 lines) into four focused modules:
- test_settings_update.py: update/start-update operations
- test_settings_save.py: save settings, validation, isolation, API keys
- test_settings_import.py: import/export operations
- test_settings_blueprint.py: logs, health, misc, helpers (remainder)

Split test_weather.py (~1000 lines) into three focused modules:
- test_weather_providers.py: OpenWeatherMap/OpenMeteo API mocking
- test_weather_validation.py: input validation (location, units, API key)
- test_weather.py: core rendering/integration tests (remainder)

No test logic changed. Test count verified: 2045 before and after.
