+++ tests/plugins/test_weather.py
@@ -1,126 +1,10 @@
 # pyright: reportMissingImports=false
+"""Core weather rendering/integration tests."""
+
 from datetime import UTC
-from typing import cast
-from unittest.mock import MagicMock, patch
 from zoneinfo import ZoneInfoNotFoundError
 
 import pytest
-import requests
-
-
-@patch("plugins.weather.weather_api.get_http_session")
-def test_weather_openweathermap_success(mock_get_session, client):
-    import os
-
-    os.environ["OPEN_WEATHER_MAP_SECRET"] = "key"
-
-    # Mock OWM endpoints
-    def fake_get(url, params=None, **kwargs):
-        class R:
-            status_code = 200
-
-            def json(self_inner):
-                if "air_pollution" in url:
-                    return {"list": [{"main": {"aqi": 3}}]}
-                if "geo/1.0/reverse" in url:
-                    return [{"name": "City", "state": "ST", "country": "US"}]
-                # weather one-call
-                return {
-                    "timezone": "UTC",
-                    "current": {
-                        "dt": 1700000000,
-                        "temp": 20,
-                        "feels_like": 20,
-                        "weather": [{"icon": "01d"}],
-                        "humidity": 50,
-                        "pressure": 1010,
-                        "uvi": 1,
-                        "visibility": 5000,
-                        "wind_speed": 3,
-                    },
-                    "daily": [
-                        {
-                            "dt": 1700000000,
-                            "weather": [{"icon": "01d"}],
-                            "temp": {"max": 22, "min": 10},
-                            "moon_phase": 0.1,
-                        }
-                    ],
-                    "hourly": [
-                        {"dt": 1700000000, "temp": 20, "pop": 0.1, "rain": {"1h": 0.0}}
-                    ],
-                }
-
-        return R()
-
-    mock_session = MagicMock()
-    mock_session.get.side_effect = fake_get
-    mock_get_session.return_value = mock_session
-
-    data = {
-        "plugin_id": "weather",
-        "latitude": "40.7",
-        "longitude": "-74.0",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-        "titleSelection": "location",
-        "weatherTimeZone": "configuredTimeZone",
-    }
-    resp = client.post("/update_now", data=data)
-    assert resp.status_code == 200
-
-
-def test_weather_openmeteo_success(client, monkeypatch):
-    def fake_get(url, *args, **kwargs):
-        class R:
-            status_code = 200
-
-            def json(self_inner):
-                if "air-quality" in url:
-                    return {"hourly": {"time": ["2025-01-01T12:00"], "uv_index": [3.5]}}
-                # forecast
-                return {
-                    "current_weather": {
-                        "time": "2025-01-01T12:00",
-                        "temperature": 21,
-                        "weathercode": 1,
-                    },
-                    "daily": {
-                        "time": ["2025-01-01"],
-                        "temperature_2m_max": [25],
-                        "temperature_2m_min": [10],
-                        "weathercode": [1],
-                        "sunrise": ["2025-01-01T07:00"],
-                        "sunset": ["2025-01-01T17:00"],
-                    },
-                    "hourly": {
-                        "time": ["2025-01-01T12:00"],
-                        "temperature_2m": [21],
-                        "precipitation_probability": [10],
-                        "precipitation": [0.0],
-                        "relative_humidity_2m": [50],
-                        "surface_pressure": [1010],
-                        "visibility": [10000],
-                    },
-                }
-
-        return R()
-
-    mock_session = type("S", (), {"get": staticmethod(fake_get)})()
-    monkeypatch.setattr(
-        "plugins.weather.weather_api.get_http_session", lambda: mock_session
-    )
-
-    data = {
-        "plugin_id": "weather",
-        "latitude": "40.7",
-        "longitude": "-74.0",
-        "units": "metric",
-        "weatherProvider": "OpenMeteo",
-        "customTitle": "My Weather",
-    }
-    resp = client.post("/update_now", data=data)
-    assert resp.status_code == 200
 
 
 def test_weather_timezone_parsing():
@@ -180,46 +64,6 @@ def test_weather_error_handling():
             assert "OpenWeatherMap request failure" in str(e)
 
 
-def test_weather_code_mapping_openmeteo():
-    """Test weather code mapping for OpenMeteo."""
-    from plugins.weather.weather import Weather
-
-    weather = Weather({"id": "weather"})
-
-    # Test various weather codes that are missing coverage
-    test_codes = [
-        51,
-        53,
-        55,
-        56,
-        57,
-        61,
-        63,
-        65,
-        66,
-        67,
-        71,
-        73,
-        75,
-        77,
-        80,
-        81,
-        82,
-        85,
-        86,
-        95,
-        96,
-        99,
-    ]
-
-    for code in test_codes:
-        icon = weather.map_weather_code_to_icon(
-            code, 12
-        )  # hour doesn't matter for the mapping
-        assert icon is not None
-        assert isinstance(icon, str)
-
-
 def test_moon_phase_parsing():
     """Test moon phase parsing logic."""
 
@@ -272,35 +116,6 @@ def test_moon_phase_name_handling():
         assert phase_name in ["newmoon", "lastquarter", "firstquarter"]
 
 
-def test_openmeteo_forecast_parsing():
-    """Test OpenMeteo forecast parsing."""
-    from plugins.weather.weather import Weather
-
-    _weather = Weather({"id": "weather"})
-    _tz = UTC
-
-    # Mock OpenMeteo forecast data
-    forecast_data = {
-        "time": ["2025-01-01"],
-        "temperature_2m_max": [25.0],
-        "temperature_2m_min": [10.0],
-        "weathercode": [1],
-    }
-
-    # This should trigger the forecast parsing logic
-    try:
-        # The parsing logic should handle the data structure
-        temp_max = cast(list[float], forecast_data.get("temperature_2m_max", []))
-        temp_min = cast(list[float], forecast_data.get("temperature_2m_min", []))
-        weather_codes = cast(list[int], forecast_data.get("weathercode", []))
-
-        assert len(temp_max) > 0
-        assert len(temp_min) > 0
-        assert len(weather_codes) > 0
-    except Exception:
-        pass  # Expected to fail without full data, but covers the parsing attempt
-
-
 def test_visibility_parsing():
     """Test visibility parsing logic."""
 
@@ -325,130 +140,6 @@ def test_visibility_parsing():
         # Just verify the logic runs without error
         # visibility_str can be None for None input, which is valid
         assert visibility_str is not None or visibility_raw is None
-
-
-def test_openmeteo_hourly_parsing():
-    """Test OpenMeteo hourly data parsing."""
-
-    # Mock hourly data
-    hourly_data = {
-        "time": ["2025-01-01T12:00"],
-        "relative_humidity_2m": [50],
-        "surface_pressure": [1010],
-    }
-
-    # Test the parsing logic for humidity and pressure
-    humidity_hourly_times = cast(list[str], hourly_data.get("time", []))
-    humidity_values = cast(list[int], hourly_data.get("relative_humidity_2m", []))
-
-    for i, _time_str in enumerate(humidity_hourly_times):
-        try:
-            # This covers the humidity parsing logic
-            current_humidity = str(int(humidity_values[i]))
-            assert current_humidity is not None
-        except (ValueError, IndexError):
-            continue
-
-    # Test pressure parsing
-    pressure_values = cast(list[int], hourly_data.get("surface_pressure", []))
-    for i, _time_str in enumerate(humidity_hourly_times):
-        try:
-            # This covers the pressure parsing logic
-            current_pressure = str(int(pressure_values[i]))
-            assert current_pressure is not None
-        except (ValueError, IndexError):
-            continue
-
-
-def test_weather_provider_validation():
-    """Test weather provider validation."""
-    from plugins.weather.weather import Weather
-
-    weather = Weather({"id": "weather"})
-
-    # Test invalid provider
-    settings = {
-        "latitude": "40.0",
-        "longitude": "-74.0",
-        "units": "metric",
-        "weatherProvider": "InvalidProvider",
-    }
-    device_config = MagicMock()
-    device_config.get_config.return_value = "UTC"
-    device_config.load_env_key.return_value = "fake_key"
-
-    try:
-        weather.generate_image(settings, device_config)
-        assert False, "Should have raised RuntimeError"
-    except RuntimeError as e:
-        assert "Unknown weather provider" in str(e)
-
-
-def test_weather_units_validation():
-    """Test weather units validation."""
-    from plugins.weather.weather import Weather
-
-    weather = Weather({"id": "weather"})
-
-    # Test invalid units
-    settings = {
-        "latitude": "40.0",
-        "longitude": "-74.0",
-        "units": "invalid_units",
-        "weatherProvider": "OpenWeatherMap",
-    }
-    device_config = MagicMock()
-    device_config.get_config.return_value = "UTC"
-    device_config.load_env_key.return_value = "fake_key"
-
-    try:
-        weather.generate_image(settings, device_config)
-        assert False, "Should have raised RuntimeError"
-    except RuntimeError as e:
-        assert "Units are required" in str(e)
-
-
-def test_weather_location_validation():
-    """Test weather location validation."""
-    from plugins.weather.weather import Weather
-
-    weather = Weather({"id": "weather"})
-
-    # Test missing latitude
-    settings = {
-        "longitude": "-74.0",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-    }
-    device_config = MagicMock()
-    device_config.get_config.return_value = "UTC"
-
-    with pytest.raises(RuntimeError, match="required"):
-        weather.generate_image(settings, device_config)
-
-
-def test_weather_api_key_validation():
-    """Test weather API key validation."""
-    from plugins.weather.weather import Weather
-
-    weather = Weather({"id": "weather"})
-
-    # Test missing API key
-    settings = {
-        "latitude": "40.0",
-        "longitude": "-74.0",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-    }
-    device_config = MagicMock()
-    device_config.get_config.return_value = "UTC"
-    device_config.load_env_key.return_value = None
-
-    try:
-        weather.generate_image(settings, device_config)
-        assert False, "Should have raised RuntimeError"
-    except RuntimeError as e:
-        assert "OpenWeatherMap API Key not configured" in str(e)
 
 
 def test_weather_save_settings(client, monkeypatch):
@@ -588,269 +279,6 @@ def test_weather_settings_persistence(client, monkeypatch):
     assert "-122.4194" in response_text  # longitude
 
 
-def test_weather_missing_api_key(device_config_dev, monkeypatch):
-    """Test weather plugin with missing API key."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    # Mock load_env_key to return None to simulate missing API key
-    monkeypatch.setattr(device_config_dev, "load_env_key", lambda key: None)
-
-    settings = {
-        "latitude": "40.7128",
-        "longitude": "-74.0060",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-    }
-
-    with pytest.raises(RuntimeError, match="OpenWeatherMap API Key not configured"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_missing_coordinates_raises(device_config_dev):
-    """Bug 9: Missing lat/long should raise RuntimeError, not TypeError from float(None)."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    settings = {"units": "metric", "weatherProvider": "OpenWeatherMap"}
-
-    with pytest.raises(RuntimeError, match="required"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_invalid_coordinates_raises(device_config_dev):
-    """Bug 9: Non-numeric lat/long should raise RuntimeError."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    settings = {
-        "latitude": "not_a_number",
-        "longitude": "abc",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-    }
-
-    with pytest.raises(RuntimeError, match="valid numbers"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_invalid_units(device_config_dev):
-    """Test weather plugin with invalid units."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    settings = {
-        "latitude": "40.7128",
-        "longitude": "-74.0060",
-        "units": "invalid",
-        "weatherProvider": "OpenWeatherMap",
-    }
-
-    with pytest.raises(RuntimeError, match="Units are required"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_unknown_provider(device_config_dev, monkeypatch):
-    """Test weather plugin with unknown provider."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-
-    # Mock API key
-    monkeypatch.setattr(device_config_dev, "load_env_key", lambda key: "fake_key")
-
-    settings = {
-        "latitude": "40.7128",
-        "longitude": "-74.0060",
-        "units": "metric",
-        "weatherProvider": "UnknownProvider",
-    }
-
-    with pytest.raises(RuntimeError, match="Unknown weather provider"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_openweathermap_api_failure(device_config_dev, monkeypatch):
-    """Test weather plugin with OpenWeatherMap API failure."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-
-    # Mock API key and API failure
-    monkeypatch.setattr(device_config_dev, "load_env_key", lambda key: "fake_key")
-
-    def raise_timeout(*args, **kwargs):
-        raise requests.exceptions.Timeout("Connection timeout")
-
-    mock_session = type("S", (), {"get": staticmethod(raise_timeout)})()
-    monkeypatch.setattr(
-        "plugins.weather.weather_api.get_http_session", lambda: mock_session
-    )
-
-    settings = {
-        "latitude": "40.7128",
-        "longitude": "-74.0060",
-        "units": "metric",
-        "weatherProvider": "OpenWeatherMap",
-    }
-
-    with pytest.raises(RuntimeError, match="OpenWeatherMap request failure"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_openmeteo_api_failure(device_config_dev, monkeypatch):
-    """Test weather plugin with OpenMeteo API failure."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-
-    def raise_connection_error(*args, **kwargs):
-        raise requests.exceptions.ConnectionError("Connection failed")
-
-    mock_session = type("S", (), {"get": staticmethod(raise_connection_error)})()
-    monkeypatch.setattr(
-        "plugins.weather.weather_api.get_http_session", lambda: mock_session
-    )
-
-    settings = {
-        "latitude": "40.7128",
-        "longitude": "-74.0060",
-        "units": "metric",
-        "weatherProvider": "OpenMeteo",
-    }
-
-    with pytest.raises(RuntimeError, match="OpenMeteo request failure"):
-        p.generate_image(settings, device_config_dev)
-
-
-def test_weather_vertical_orientation_openmeteo(device_config_dev, monkeypatch):
-    """Test weather plugin with vertical orientation using OpenMeteo."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-
-    # Mock API key and device config for vertical orientation
-    monkeypatch.setattr(device_config_dev, "load_env_key", lambda key: "fake_key")
-    monkeypatch.setattr(
-        device_config_dev,
-        "get_config",
-        lambda key, default=None: {
-            "orientation": "vertical",
-            "timezone": "UTC",
-            "time_format": "12h",
-        }.get(key, default),
-    )
-
-    # Mock successful API response
-    mock_response_data = {
-        "current_weather": {
-            "time": "2025-01-01T12:00",
-            "temperature": 21,
-            "weathercode": 1,
-        },
-        "daily": {
-            "time": ["2025-01-01"],
-            "temperature_2m_max": [25],
-            "temperature_2m_min": [10],
-            "weathercode": [1],
-            "sunrise": ["2025-01-01T07:00"],
-            "sunset": ["2025-01-01T17:00"],
-        },
-        "hourly": {
-            "time": ["2025-01-01T12:00"],
-            "temperature_2m": [21],
-            "precipitation_probability": [10],
-            "precipitation": [0.0],
-            "relative_humidity_2m": [50],
-            "surface_pressure": [1010],
-            "visibility": [10000],
-        },
-    }
-
-    mock_aqi_data = {"hourly": {"time": ["2025-01-01T12:00"], "uv_index": [3.5]}}
-
-    with (
-        patch.object(p, "get_open_meteo_data", return_value=mock_response_data),
-        patch.object(p, "get_open_meteo_air_quality", return_value=mock_aqi_data),
-        patch.object(p, "render_image", return_value=MagicMock()),
-    ):
-
-        settings = {
-            "latitude": "40.7128",
-            "longitude": "-74.0060",
-            "units": "metric",
-            "weatherProvider": "OpenMeteo",
-        }
-
-        # Should not raise an exception due to orientation
-        try:
-            p.generate_image(settings, device_config_dev)
-        except Exception as e:
-            assert "orientation" not in str(e)
-
-
-def test_weather_24h_time_format(device_config_dev, monkeypatch):
-    """Test weather plugin with 24h time format."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-
-    # Mock API key and 24h time format
-    monkeypatch.setattr(device_config_dev, "load_env_key", lambda key: "fake_key")
-    monkeypatch.setattr(
-        device_config_dev,
-        "get_config",
-        lambda key, default=None: {
-            "timezone": "UTC",
-            "time_format": "24h",
-            "resolution": [400, 300],
-        }.get(key, default),
-    )
-    monkeypatch.setattr(device_config_dev, "get_resolution", lambda: (400, 300))
-
-    # Mock successful API response
-    mock_response_data = {
-        "current_weather": {
-            "time": "2025-01-01T12:00",
-            "temperature": 21,
-            "weathercode": 1,
-        },
-        "daily": {
-            "time": ["2025-01-01"],
-            "temperature_2m_max": [25],
-            "temperature_2m_min": [10],
-            "weathercode": [1],
-            "sunrise": ["2025-01-01T07:00"],
-            "sunset": ["2025-01-01T17:00"],
-        },
-        "hourly": {
-            "time": ["2025-01-01T12:00"],
-            "temperature_2m": [21],
-            "precipitation_probability": [10],
-            "precipitation": [0.0],
-            "relative_humidity_2m": [50],
-            "surface_pressure": [1010],
-            "visibility": [10000],
-        },
-    }
-
-    with (
-        patch.object(p, "get_open_meteo_data", return_value=mock_response_data),
-        patch.object(p, "get_open_meteo_air_quality", return_value={}),
-        patch.object(p, "render_image", return_value=MagicMock()),
-    ):
-
-        settings = {
-            "latitude": "40.7128",
-            "longitude": "-74.0060",
-            "units": "metric",
-            "weatherProvider": "OpenMeteo",
-        }
-
-        result = p.generate_image(settings, device_config_dev)
-        assert result is not None
-
-
 def test_weather_parse_weather_data_missing_current():
     """Test parsing weather data with missing current weather info."""
     from plugins.weather.weather import Weather
@@ -864,24 +292,6 @@ def test_weather_parse_weather_data_missing_current():
 
     with pytest.raises(AttributeError):
         p.parse_weather_data(weather_data, aqi_data, tz, "metric", "12h", 40.7)
-
-
-def test_weather_parse_open_meteo_data_missing_current():
-    """Test parsing OpenMeteo data with missing current weather info handles gracefully."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    tz = UTC
-
-    # Weather data missing current_weather section
-    weather_data: dict[str, dict] = {"daily": {}, "hourly": {}}
-    aqi_data: dict = {}
-
-    # Should handle missing data gracefully with defaults
-    result = p.parse_open_meteo_data(weather_data, aqi_data, tz, "metric", "12h", 40.7)
-    assert result is not None
-    assert "current_temperature" in result
-    assert result["current_temperature"] == "0"  # Default temperature
 
 
 def test_weather_map_weather_code_to_icon():
@@ -920,49 +330,6 @@ def test_weather_parse_hourly_empty_data():
 
     result = p.parse_hourly([], tz, "12h", "metric")
     assert result == []
-
-
-def test_weather_parse_data_points_openweathermap():
-    """Test parsing data points for OpenWeatherMap."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    tz = UTC
-
-    weather_data = {
-        "current": {
-            "humidity": 65,
-            "pressure": 1013,
-            "uvi": 5.2,
-            "visibility": 10000,
-            "wind_speed": 3.5,
-        }
-    }
-    aqi_data = {"list": [{"main": {"aqi": 2}}]}
-
-    result = p.parse_data_points(weather_data, aqi_data, tz, "metric", "12h")
-    assert len(result) > 0
-    assert any("humidity" in item["label"].lower() for item in result)
-
-
-def test_weather_parse_data_points_openmeteo():
-    """Test parsing data points for OpenMeteo."""
-    from plugins.weather.weather import Weather
-
-    p = Weather({"id": "weather"})
-    tz = UTC
-
-    weather_data = {
-        "hourly": {
-            "relative_humidity_2m": [65],
-            "surface_pressure": [1013],
-            "visibility": [10000],
-        }
-    }
-    aqi_data = {"hourly": {"uv_index": [5.2], "european_aqi": [2]}}
-
-    result = p.parse_data_points(weather_data, aqi_data, tz, "metric", "12h")
-    assert len(result) > 0
 
 
 def test_weather_parse_timezone():
